Time: 14 ms (14.20%) | Memory: 7.4 MB (80.97%) - LeetSync

+++ 69-sqrtx/sqrtx.cpp
@@ -2,8 +2,8 @@ class Solution {
 public:
     int mySqrt(int x) {
       if (x <= 1) return x;
-long i = 1;
-long sq = i * i;
+unsigned int i = 1;
+unsigned int sq = i * i;
 while (x >= sq) {
 	i++;
 	sq = i * i;
